+++ taglib/trueaudio/trueaudiofile.cpp
@@ -64,7 +64,7 @@ public:
 
   const ID3v2::FrameFactory *ID3v2FrameFactory;
   long ID3v2Location;
-  unsigned int ID3v2OriginalSize;
+  long ID3v2OriginalSize;
 
   long ID3v1Location;
 
@@ -169,7 +169,7 @@ bool TrueAudio::File::save()
     insert(data, d->ID3v2Location, d->ID3v2OriginalSize);
 
     if(d->ID3v1Location >= 0)
-      d->ID3v1Location += (data.size() - d->ID3v2OriginalSize);
+      d->ID3v1Location += (static_cast<long>(data.size()) - d->ID3v2OriginalSize);
 
     d->ID3v2OriginalSize = data.size();
   }
